pyproject.toml
[project]
name = "pythonclient"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Shlule", email = "93739679+Shlule@users.noreply.github.com" }
]
dependencies = [
    "python-socketio[asyncio_client]>=5.11.1",
    "asyncio>=3.4.3",
    "logzero>=1.7.0",
    "argparse>=1.4.0",
    "dacite>=1.8.1",
    "jsondiff>=2.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.3.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pythonclient"]
